Move extract utils from plan to root utils file
